% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nimble.R
\name{find_sampler_indices_fast}
\alias{find_sampler_indices_fast}
\title{Find sampler indices (fast)}
\usage{
find_sampler_indices_fast(conf, nodes)
}
\arguments{
\item{conf}{A \code{\link[nimble]{MCMCconf}} object from the \pkg{nimble} package.}

\item{nodes}{A parameter name (character string).}
}
\value{
An integer vector of indices into \code{conf$getSamplers()}.
}
\description{
This function provides a fast alternative to \code{conf$findSamplersOnNodes()},
which can be slow for large target samplers.
}
