% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior.R
\name{get_post_summary}
\alias{get_post_summary}
\title{Extract posterior summary of parameters from a model-fit object.}
\usage{
get_post_summary(
  fit,
  parameter = c("z", "pi", "phi", "theta", "psi", "alpha", "beta", "gamma",
    "alpha_shared", "beta_shared", "gamma_shared", "Mu", "sigma", "rho")
)
}
\arguments{
\item{fit}{An \code{occumbFit} object.}

\item{parameter}{TODO: to be added.}
}
\value{
TODO: to be added.
}
\description{
TODO: to be added.
}
\details{
TODO: to be added.
}
\examples{
# to be added
}
