% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof}
\alias{gof}
\title{Goodness-of-fit assessment.}
\usage{
gof(fit, data, plot = TRUE)
}
\arguments{
\item{fit}{An \code{occumbFit} object.}

\item{data}{An \code{occumbData} object used for the model fitting.}

\item{plot}{Logical, determine if draw scatter plots of the fit statistics.}
}
\description{
\code{gof()} computes some fit statistics and their Bayesian \eqn{p}-values from the model fitting results and the original data.
}
\details{
TODO: Add details.
- This function conducts the procedure of gof assessment described in (ref).
- Two statistics: deviance and Freeman_Tukey statistics.
- Bayesian p-value: I(obs < rep) / N.
Extreme values may indicate inadequate model fit.
- References for gof stats and Bayesian p-values.
}
\section{References}{

TODO: Add references.
}

\examples{
\dontrun{
# Generate the smallest random dataset (2 species * 2 sites * 2 reps)
I <- 2 # Number of species
J <- 2 # Number of sites
K <- 2 # Number of replicates
data <- occumbData(
    y = array(sample.int(I * J * K), dim = c(I, J, K)),
    spec_cov = list(cov1 = rnorm(I)),
    site_cov = list(cov2 = rnorm(J),
                    cov3 = factor(1:J)),
    repl_cov = list(cov4 = matrix(rnorm(J * K), J, K)))

# Fitting a null model
fit <- occumb(data = data)

# Goodness-of-fit assessment
gof_result <- gof(fit, data)
}
}
