% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occumbData.R
\name{occumbData}
\alias{occumbData}
\title{Constructor for occumbData data class.}
\usage{
occumbData(y, spec_cov = NULL, site_cov = NULL, repl_cov = NULL)
}
\arguments{
\item{y}{A 3-D array of sequence read counts (integer values).
Dimensions are ordered by species, site, and replicate.
Data for missing replicates must be represented by zero vectors.
NAs are not allowed.}

\item{spec_cov}{A named list of species covariates.
Each element must be a vector of numeric, factor, or
character whose length is equal to dim(y)[1] (i.e., the
number of species). Covariates in character are automatically
converted to factors. NAs are not allowed.}

\item{site_cov}{A named list of site covariates.
Each element must be a vector of numeric, factor, or
character whose length is equal to dim(y)[2] (i.e., the
number of sites). Covariates in character are automatically
converted to factors. NAs are not allowed.}

\item{repl_cov}{A named list of replicate covariates.
Each element must be a matrix of numeric, factor, or
character whose dimension is equal to dim(y)[2:3] (i.e.,
the number of sites * number of replicates). Covariates in
character are automatically converted to factors. NAs are
not allowed.}
}
\value{
An S4 object of the \code{occumbData} class.
}
\description{
\code{occumbData} creates a data list compatible with the model-fitting
function \code{occumb}.
}
\section{Details}{

The element names for spec_cov, site_cov, and repl_cov must all be unique.
}

\examples{
# Generate a small, random dataset
data <- occumbData(
    y = array(sample.int(8), dim = rep(2, 3)),
    spec_cov = list(cov1 = rnorm(2)),
    site_cov = list(cov2 = rnorm(2),
                    cov3 = factor(1:2)))
}
