% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,occumbFit-method}
\alias{plot,occumbFit-method}
\title{Plot method for occumbFit class.}
\usage{
\S4method{plot}{occumbFit}(x, y, ...)
}
\arguments{
\item{x}{An occumbFit object.}

\item{y}{Not used.}

\item{...}{Other arguments passed to the plot method for
\code{jagsUI} class.}
}
\description{
Plot method for occumbFit class.
}
