% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,occumbGof-method}
\alias{plot,occumbGof-method}
\title{Plot method for occumbGof class.}
\usage{
\S4method{plot}{occumbGof}(x, y = NULL, ...)
}
\arguments{
\item{x}{An \code{occumbGof} object.}

\item{y}{\code{NULL}}

\item{...}{Additional arguments passed to the default \code{plot} method.}
}
\description{
Draws a scatter plot of fit statistics.
}
