% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print,occumbFit-method}
\alias{print,occumbFit-method}
\title{Print method for occumbFit class.}
\usage{
\S4method{print}{occumbFit}(x, ...)
}
\arguments{
\item{x}{An occumbFit object.}

\item{...}{Other arguments passed to the print method for
\code{jagsUI} class.}
}
\description{
Print method for occumbFit class.
}
