% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,occumbData-method}
\alias{summary,occumbData-method}
\title{Summary method for occumbData class.}
\usage{
\S4method{summary}{occumbData}(object)
}
\arguments{
\item{object}{An \code{occumbData} object.}
}
\description{
Summarizes dataset stored in an \code{occumbData} object.
}
