% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,occumbFit-method}
\alias{summary,occumbFit-method}
\title{Summary method for occumbFit class.}
\usage{
\S4method{summary}{occumbFit}(object)
}
\arguments{
\item{object}{An \code{occumbFit} object.}
}
\description{
Summarizes model fitting result stored in an \code{occumbFit} object.
}
