% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{eval_util_R}
\alias{eval_util_R}
\title{Expected utility for regional species diversity assessments.}
\usage{
eval_util_R(settings, fit, N_rep = 1, cores = parallel::detectCores())
}
\arguments{
\item{settings}{A data frame that specifies a set of conditions under which
the utility is evaluated. It must include a column named \code{J}, \code{K}, and \code{N},
which specifies the number of sites, the number of replicates per site, and
the sequencing depth per replicate, respectively.
\code{J}, \code{K}, and \code{N} must be numeric vectors greater than 0. When \code{J} and \code{K}
contains a decimal, the decimal part is discarded and treated as an integer.
Additional columns are ignored but may be included.}

\item{fit}{An \code{occumbFit} object containing a posterior sample of the
relevant parameters.}

\item{N_rep}{Controls the sample size for Monte Carlo integration.
The integral is evaluated using a total of \code{N_sample * N_rep} random samples,
where \code{N_sample} is the size of the MCMC sample provided as \code{fit}.}

\item{cores}{The number of cores to use for parallelization.}
}
\value{
A data frame with a column named \code{Utility} in which the estimates of
expected utility are stored. This is obtained by adding the \code{Utility} column
to the data frame provided in the \code{settings} argument.
}
\description{
\code{eval_util_R()} evaluates the expected utility for the regional
species diversity assessment using Monte Carlo integration.
}
\details{
The utility of the regional species diversity assessment can be defined as
the number of species in the region of interest expected to be detected
(Fukaya et al. 2022). \code{eval_util_R()} evaluates this utility for the region
modeled in the \code{occumbFit} object for the combination of \code{J}, \code{K}, and \code{N}
values specified in the \code{conditions} argument.
Such evaluations can be used to balance \code{J}, \code{K}, and \code{N} to maximize the
utility under a constant budget (possible combinations of \code{J}, \code{K}, and \code{N}
under a budget level are easily obtained using \code{x}; see the example below).
It is also possible to examine how the utility varies with different \code{J},
\code{K}, and \code{N} values without setting a budget level, which may be useful to know
the satisfactory level of \code{J}, \code{K}, and \code{N} from a purely technical point of
view.

The expected utility is defined as the expected value of the conditional
utility of the form:
\deqn{U(J, K, N \mid \boldsymbol{r}, \boldsymbol{u}) = \sum_{i = 1}^{I}\left\{1 - \prod_{j = 1}^{J}\prod_{k = 1}^{K}\left(1 - \frac{u_{ijk}r_{ijk}}{\sum_{m = 1}^{I}u_{mjk}r_{mjk}} \right)^N \right\}}{U(J, K, N | r, u) = sum_{i}((1 - \prod_{j}\prod_{k}(1 - (u[i, j, k] * r[i, j, k])/sum(u[, j, k] * r[, j, k])))^N)}
where \eqn{u_{ijk}}{u[i, j, k]} is a latent indicator variable representing
the inclusion of the sequence of species \eqn{i}{i} in replicate \eqn{k}{k}
at site \eqn{j}{j} and \eqn{r_{ijk}}{r[i, j, k]} is a latent variable that
is proportional to the relative frequency of the sequence of species
\eqn{i}{i}, conditional on its presence in replicate \eqn{k}{k} at site
\eqn{j}{j} (Fukaya et al. 2022).
Expectations are taken with respect to the posterior predictive distributions
of \eqn{\boldsymbol{r} = \{r_{ijk}\}}{r} and
\eqn{\boldsymbol{u} = \{u_{ijk}\}}{u}, which are evaluated numerically by
Monte Carlo integration using MCMC samples in the \code{occumbFit} object. Higher
approximation accuracy can be obtained by increasing the value of \code{N_rep}.

The expected utility is evaluated assuming that all sites and replicates are
homogeneous in the sense that the model parameters are constant across sites
and replicates. For this reason, in the current version, \code{eval_util_R()}
cannot be applied if the supplied \code{occumbFit} object contains a model with
site or replicate covariates.

Monte Carlo integration is executed in parallel on multiple CPU cores where
the \code{cores} argument controls the degree of parallelization. By default, all
cores available in the user's environment are used.
}
\section{References}{

K. Fukaya, N. I. Kondo, S. S. Matsuzaki and T. Kadoya (2022)
Multispecies site occupancy modelling and study design for spatially
replicated environmental DNA metabarcoding. \emph{Methods in Ecology
and Evolution} \strong{13}:183--193.
\url{https://doi.org/10.1111/2041-210X.13732}
}

