% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{loglik}
\alias{loglik}
\title{Pointwise log-likelihood of the fitted model.}
\usage{
loglik(fit)
}
\arguments{
\item{fit}{An \code{occumbFit} object.}
}
\value{
An \eqn{M \times L} matrix, where \eqn{M} is the size of the
posterior sample and \eqn{L} is the number of replicates.
}
\description{
\code{loglik()} extracts the pointwise log-likelihood matrix from
an \code{occumbFit} object.
}
\details{
The pointwise log-likelihood is the log-likelihood of each data point
evaluated with the posterior samples of parameter values. It can be used to
obtain criteria for model comparisons, including WAIC and PSIS-LOO. These
criteria can be conveniently calculated by applying functions from the
\href{https://cran.r-project.org/web/packages/loo/index.html}{\code{loo}} package
to the output of this function: see Vehtari et al. (2017) and the
documentation of the \code{loo} package for details.
}
\section{References}{

A. Vehtari, A. Gelman, J. Gabry (2017) Practical Bayesian model evaluation
using leave-one-out cross-validation and WAIC.
\emph{Statistics and Computing}, \strong{27}, 1413--1432.
\url{https://doi.org/10.1007/s11222-016-9696-4}
}

\examples{
\dontrun{
# Generate the smallest random dataset (2 species * 2 sites * 2 reps)
I <- 2 # Number of species
J <- 2 # Number of sites
K <- 2 # Number of replicates
data <- occumbData(
    y = array(sample.int(I * J * K), dim = c(I, J, K)),
    spec_cov = list(cov1 = rnorm(I)),
    site_cov = list(cov2 = rnorm(J),
                    cov3 = factor(1:J)),
    repl_cov = list(cov4 = matrix(rnorm(J * K), J, K)))

# Fitting a null model
fit <- occumb(data = data)

# Extract log-likelihood
ll <- loglik(fit)  # could be passed to loo() or waic() in the loo package
}
}
