% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occumb.R
\name{occumb}
\alias{occumb}
\title{Model-fitting function.}
\usage{
occumb(
  phi_formula = ~1,
  theta_formula = ~1,
  psi_formula = ~1,
  data,
  prior_prec = 0.001,
  prior_ulim = 1000,
  n.chains = 6,
  n.adapt = 1000,
  n.burnin = 30000,
  n.thin = 500,
  n.iter = 500 * n.thin + n.burnin,
  parallel = TRUE
)
}
\arguments{
\item{phi_formula}{A right-hand side formula describing covariates of
sequence relative dominance.}

\item{theta_formula}{A right-hand side formula describing covariates of
sequence capture probability.}

\item{psi_formula}{A right-hand side formula describing covariates of
occupancy probability.}

\item{data}{occumbData class object.}

\item{prior_prec}{Precision of the normal prior distribution for the
community-level average of species-specific parameters.}

\item{prior_ulim}{Upper limit of the uniform prior distribution for the
standard deviation of species-specific parameters.}

\item{n.chains}{Number of Markov chains to run.
Passed to the \code{n.chains} argument of jagsUI's \code{jags()}
function.}

\item{n.adapt}{Number of iterations to run in the JAGS adaptive phase.
Passed to the \code{n.adapt} argument of jagsUI's \code{jags()}
function.
See the document of jagsUI's \code{jags()} function for details.}

\item{n.burnin}{Number of iterations at the beginning of the chain to discard.
Passed to the \code{n.burnin} argument of jagsUI's \code{jags()}
function.
See the document of jagsUI's \code{jags()} function for details.}

\item{n.thin}{Thinning rate. Must be a positive integer.
Passed to the \code{n.thin} argument of jagsUI's \code{jags()}
function.}

\item{n.iter}{Total number of iterations per chain (including burn-in).
Passed to the \code{n.iter} argument of jagsUI's \code{jags()}
function.}

\item{parallel}{If TRUE, run MCMC chains in parallel on multiple CPU cores.
Passed to the \code{parallel} argument of jagsUI's \code{jags()}
function.}
}
\value{
An S4 object of the \code{occumbFit} class containing the results of
model fitting.
}
\description{
\code{occumb()} fits a specified site-occupancy model to a dataset and returns
a model-fit object containing posterior samples.
}
\section{Details}{

\code{occumb()} allows the fitting of a range of site occupancy models
for sequence read counts that includes covariates at different levels
of the data generation process, via the specification of the model
formula in \code{phi_formula}, \code{theta_formula}, and
\code{psi_formula.}
The specified model and dataset are passed to JAGS via the
interface provided by the jagsUI package, where Markov chain
Monte Carlo methods are used to obtain posterior samples of model
parameters.\preformatted{ For the resulting object, the following methods and functions provided
 by the jagsUI package can be applied:

 plot(); print(); summary()

 Currently, occumb() supports covariate modeling only for
 psi_formula in which site covariates can be incorporated.
}
}

