% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occumb.R
\name{occumb}
\alias{occumb}
\title{Model-fitting function.}
\usage{
occumb(
  formula_phi = ~1,
  formula_theta = ~1,
  formula_psi = ~1,
  formula_phi_shared = ~1,
  formula_theta_shared = ~1,
  formula_psi_shared = ~1,
  prior_prec = 1e-04,
  prior_ulim = 10000,
  data,
  n.chains = 4,
  n.adapt = NULL,
  n.burnin = 10000,
  n.thin = 10,
  n.iter = 20000,
  parallel = FALSE
)
}
\arguments{
\item{formula_phi}{A right-hand side formula describing species-specific
effects of sequence relative dominance (\eqn{\phi}).}

\item{formula_theta}{A right-hand side formula describing species-specific
effects of sequence capture probability (\eqn{\theta}).}

\item{formula_psi}{A right-hand side formula describing species-specific
effects of occupancy probability (\eqn{\psi}).}

\item{formula_phi_shared}{A right-hand side formula describing effects of
sequence relative dominance (\eqn{\phi}) that are common across
species. The intercept term is ignored (see Details).}

\item{formula_theta_shared}{A right-hand side formula describing effects of
sequence capture probability (\eqn{\theta}) that are common across
species.
The intercept term is ignored (see Details).}

\item{formula_psi_shared}{A right-hand side formula describing effects of
occupancy probability (\eqn{\psi}) that are common across species.
The intercept term is ignored (see Details).}

\item{prior_prec}{Precision of the normal prior distribution for the
community-level average of species-specific parameters and effects
common across species.}

\item{prior_ulim}{Upper limit of the uniform prior distribution for the
standard deviation of species-specific parameters.}

\item{data}{A dataset supplied as an \code{\link{occumbData}} class object.}

\item{n.chains}{Number of Markov chains to run.}

\item{n.adapt}{Number of iterations to run in the JAGS adaptive phase.}

\item{n.burnin}{Number of iterations at the beginning of the chain to discard.}

\item{n.thin}{Thinning rate. Must be a positive integer.}

\item{n.iter}{Total number of iterations per chain (including burn-in).}

\item{parallel}{If TRUE, run MCMC chains in parallel on multiple CPU cores.}
}
\value{
An S4 object of the \code{occumbFit} class containing the results of
model fitting.
}
\description{
\code{occumb()} fits the multispecies site-occupancy model for eDNA
metabarcoding (Fukaya et al. 2021) and returns a model-fit object containing
posterior samples.
}
\details{
\code{occumb()} allows the fitting of a range of the multispecies site
occupancy models including covariates at different levels of the data
generation process.
The most general form of the model can be written as follows:

Sequence read counts:
\deqn{(y_{1jk}, ..., y_{Ijk}) \sim \textrm{Multinomial}((\pi_{1jk}, ...,  \pi_{Ijk}), N_{jk}),}{(y_{1jk}, ..., y_{Ijk}) ~ Multinomial((pi_{1jk}, ..., pi_{Ijk}), N_{jk}),}
\deqn{\pi_{ijk} = \frac{u_{ijk}r_{ijk}}{\sum_m u_{mjk}r_{mjk}},}{pi_{ijk} = (u_{ijk} * r_{ijk}) / (\sum_m (u_{mjk} * r_{mjk})),}

Relative frequency of species sequences:
\deqn{r_{ijk} \sim \textrm{Gamma}(\phi_{ijk}, 1),}{r_{ijk} ~ Gamma(phi_{ijk}, 1),}

Capture of species sequences:
\deqn{u_{ijk} \sim \textrm{Bernoulli}(z_{ij}\theta_{ijk}),}{u_{ijk} ~ Bernoulli(z_{ij} * theta_{ijk}),}

Site occupancy of species:
\deqn{z_{ij} \sim \textrm{Bernoulli}(\psi_{ij}),}{z_{ij} ~ Bernoulli(psi_{ij}),}
where the variations of \eqn{\phi}{phi}, \eqn{\theta}{theta}, and
\eqn{\psi}{psi} are modeled by specifying model formulas in
\code{formula_phi}, \code{formula_theta}, \code{formula_psi},
\code{formula_phi_shared}, \code{formula_theta_shared}, and
\code{formula_psi_shared.}

Each parameter may have species-specific effects and effects that are common
across species, where the former is specified by \code{formula_phi},
\code{formula_theta}, and \code{formula_psi}, while
\code{formula_phi_shared}, \code{formula_theta_shared}, and
\code{formula_psi_shared} specify the latter.
Because species-specific intercepts are specified by default, the intercept
term in the \code{formula_phi_shared}, \code{formula_theta_shared}, and
\code{formula_psi_shared} are always ignored.
The covariate terms must be found in the names of the list elements stored
in \code{spec_cov}, \code{site_cov}, or \code{repl_cov} slots in the dataset
object provided with the \code{data} argument.
Covariates are modeled using the log link function for \eqn{\phi}{phi}
and logit link function for \eqn{\theta}{theta} and \eqn{\psi.}{psi.}

The two arguments, \code{prior_prec} and \code{prior_ulim}, control the
prior distribution of parameters. For the community-level average of
species-specific effects and for effects common across species, a normal
prior distribution with mean 0 and precision (i.e., the inverse of the
variance) \code{prior_prec} is specified. For the standard deviation of
species-specific effects, a uniform prior distribution with a lower limit of
0 and an upper limit of \code{prior_ulim} is specified. For the correlation
coefficient of species-specific effects, a uniform prior distribution in the
range of \eqn{-1} to 1 is specified by default.

The \code{data} argument requires a dataset object generated using
\code{ocumbData()}: see the document of \code{\link{occumbData}()}.

The model is fit via the \code{\link[jagsUI]{jags}()} function of the
jagsUI package, where Markov chain Monte Carlo methods are used to
obtain posterior samples of parameters and latent variables.
Arguments \code{n.chains}, \code{n.adapt}, \code{n.burnin}, \code{n.thin},
\code{n.iter}, and \code{parallel} are passed on to the argument of the
same name in the \code{\link[jagsUI]{jags}()} function.
See the document of jagsUI's \code{\link[jagsUI]{jags}()} function for
details.
}
\section{References}{

K. Fukaya, N. I. Kondo, S. S. Matsuzaki and T. Kadoya (2021)
Multispecies site occupancy modelling and study design for spatially
replicated environmental DNA metabarcoding. \emph{Methods in Ecology
and Evolution}. https://doi.org/10.1111/2041-210X.13732
}

