% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,occumbFit-method}
\alias{plot,occumbFit-method}
\title{Plot method for occumbFit class.}
\usage{
\S4method{plot}{occumbFit}(x, y = NULL, ...)
}
\arguments{
\item{x}{An \code{occumbFit} object.}

\item{y}{\code{NULL}}

\item{...}{Additional arguments passed to the plot method for
\href{https://cran.r-project.org/package=jagsUI}{jagsUI} object.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Applies \href{https://cran.r-project.org/package=jagsUI}{jagsUI}'s
plot method to an \code{occumbFit} object to draw trace plots
and density plots of MCMC samples of model parameters.
}
