% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,occumbFit-method}
\alias{summary,occumbFit-method}
\title{Summary method for occumbFit class.}
\usage{
\S4method{summary}{occumbFit}(object, ...)
}
\arguments{
\item{object}{An occumbFit object.}

\item{...}{Other arguments passed to the summary method for
\code{jagsUI} class.}
}
\description{
Summary method for occumbFit class.
}
