% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleSummary.R
\name{summary.nimbleSummary}
\alias{summary.nimbleSummary}
\title{Summarize a nimbleSummary object}
\usage{
\method{summary}{nimbleSummary}(object, ...)
}
\arguments{
\item{object}{A \code{nimbleSummary} object.}

\item{...}{Unused.}
}
\value{
The summary matrix (invisibly). Printed for convenience.
}
\description{
\code{summary()} method for objects returned by \code{nimbleSummary()}.
}
